firstapp/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
Django==4.0.4
django-environ==0.8.1
django-modeltranslation==0.18.2
django-utils-six==2.0
frozenlist==1.3.0
graphene==3.0
graphene-django==3.0.0b7
graphql-core==3.1.7
graphql-relay==3.1.5
idna==3.3
multidict==6.0.2
Pillow==9.1.1
promise==2.3
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
telepot==12.7
text-unidecode==1.3
tzdata==2022.1
urllib3==1.26.9
yarl==1.7.2
